// src/lib/types/enums.ts
export enum IntervalE {
  MONTHLY = 'Monthly',
  YEARLY = 'Yearly'
}

export enum AuthProviderE {
  EMAIL = 'email',
  GOOGLE = 'google'
}

export enum RolesE {
  OWNER = 'OWNER',
  ADMIN = 'ADMIN',
  MEMBER = 'MEMBER'
}

export enum Actions {
  CREATE = 'create',
  READ = 'read',
  UPDATE = 'update',
  DELETE = 'delete',
  MANAGE = 'manage'
}

export enum Subjects {
  SUBSCRIPTION = 'Subscription',
  TODO = 'Todo',
  USER = 'User',
  ALL = 'all'
}